implemented on/off feature

+++ Testing/server.js
@@ -288,43 +288,47 @@ client.on('message', async message => {
             .catch(error => message.reply(`Couldn't delete messages because of: ${error}`));
     }
     else {
-        const vowels = ["a", "e", "i", "o", "u", "y"];
-        if (vowels.some(word => message.content.includes(word)) ) {
-            const text = message.content;
-            (async () => {
-                const result = await perspective.analyze(text, {attributes: ['toxicity', 'severe_toxicity', 'identity_attack', 'insult', 'profanity', 'sexually_explicit', 'threat', 'flirtation', 'attack_on_author', 'attack_on_commenter']});
-                const v1 = `${result.attributeScores.TOXICITY.summaryScore.value}`;
-                const v2 = `${result.attributeScores.SEVERE_TOXICITY.summaryScore.value}`;
-                const v3 = `${result.attributeScores.IDENTITY_ATTACK.summaryScore.value}`;
-                const v4 = `${result.attributeScores.INSULT.summaryScore.value}`;
-                const v5 = `${result.attributeScores.PROFANITY.summaryScore.value}`;
-                const v6 = `${result.attributeScores.SEXUALLY_EXPLICIT.summaryScore.value}`;
-                const v7 = `${result.attributeScores.THREAT.summaryScore.value}`;
-                const v8 = `${result.attributeScores.FLIRTATION.summaryScore.value}`;
-                const v9 = `${result.attributeScores.ATTACK_ON_AUTHOR.summaryScore.value}`;
-                const v10 = `${result.attributeScores.ATTACK_ON_COMMENTER.summaryScore.value}`;
+        if (config.on === "false") {
+            // Do nothing since the owners don't want it to work
+        } else {
+            const vowels = ["a", "e", "i", "o", "u", "y"];
+            if (vowels.some(word => message.content.includes(word))) {
+                const text = message.content;
+                (async () => {
+                    const result = await perspective.analyze(text, {attributes: ['toxicity', 'severe_toxicity', 'identity_attack', 'insult', 'profanity', 'sexually_explicit', 'threat', 'flirtation', 'attack_on_author', 'attack_on_commenter']});
+                    const v1 = `${result.attributeScores.TOXICITY.summaryScore.value}`;
+                    const v2 = `${result.attributeScores.SEVERE_TOXICITY.summaryScore.value}`;
+                    const v3 = `${result.attributeScores.IDENTITY_ATTACK.summaryScore.value}`;
+                    const v4 = `${result.attributeScores.INSULT.summaryScore.value}`;
+                    const v5 = `${result.attributeScores.PROFANITY.summaryScore.value}`;
+                    const v6 = `${result.attributeScores.SEXUALLY_EXPLICIT.summaryScore.value}`;
+                    const v7 = `${result.attributeScores.THREAT.summaryScore.value}`;
+                    const v8 = `${result.attributeScores.FLIRTATION.summaryScore.value}`;
+                    const v9 = `${result.attributeScores.ATTACK_ON_AUTHOR.summaryScore.value}`;
+                    const v10 = `${result.attributeScores.ATTACK_ON_COMMENTER.summaryScore.value}`;
 
-                if (v1 > 0.5 || v2 > 0.5 || v3 > 0.5 || v4 > 0.5 || v6 > 0.5 || v7 > 0.5 || v8 > 0.5 || v9 > 0.5 || v10 > 0.5) {
-                    sb.clear();
-                    if (v1 > 0.5) sb.append('**Toxicity**  ');
-                    if (v2 > 0.5) sb.append('**Severe Toxicity**  ');
-                    if (v3 > 0.5) sb.append('**Identity Attack**  ');
-                    if (v4 > 0.5) sb.append('**Insult**  ');
-                    if (v5 > 0.5) sb.append('**Profanity**  ');
-                    if (v6 > 0.5) sb.append('**Sexually Explicit**  ');
-                    if (v7 > 0.5) sb.append('**Threat**  ');
-                    if (v8 > 0.5) sb.append('**Flirtation**  ');
-                    if (v9 > 0.5) sb.append('**Attack on Author**  ');
-                    if (v10 > 0.5) sb.append('**Attack on Commenter**  ');
+                    if (v1 > 0.5 || v2 > 0.5 || v3 > 0.5 || v4 > 0.5 || v6 > 0.5 || v7 > 0.5 || v8 > 0.5 || v9 > 0.5 || v10 > 0.5) {
+                        sb.clear();
+                        if (v1 > 0.5) sb.append('**Toxicity**  ');
+                        if (v2 > 0.5) sb.append('**Severe Toxicity**  ');
+                        if (v3 > 0.5) sb.append('**Identity Attack**  ');
+                        if (v4 > 0.5) sb.append('**Insult**  ');
+                        if (v5 > 0.5) sb.append('**Profanity**  ');
+                        if (v6 > 0.5) sb.append('**Sexually Explicit**  ');
+                        if (v7 > 0.5) sb.append('**Threat**  ');
+                        if (v8 > 0.5) sb.append('**Flirtation**  ');
+                        if (v9 > 0.5) sb.append('**Attack on Author**  ');
+                        if (v10 > 0.5) sb.append('**Attack on Commenter**  ');
 
-                    message.delete();
-                    message.reply(`your message was deleted for the following reasons: ${sb.toString()}`);
-                }
+                        message.delete();
+                        message.reply(`your message was deleted for the following reasons: ${sb.toString()}`);
+                    }
 
-            })();
-        }
-        else {
-            //no u
+                })();
+            }
+            else {
+                //no u
+            }
         }
     }
 });
